move constant into query_constants.py
