+++ san/get.py
@@ -1,6 +1,5 @@
 import san.sanbase_graphql
-from san.query_constants import DEPRECATED_QUERIES, CUSTOM_QUERIES, NO_SLUG_QUERIES
-from san.sanbase_graphql_helper import QUERY_MAPPING
+from san.query_constants import DEPRECATED_QUERIES, CUSTOM_QUERIES, NO_SLUG_QUERIES, QUERY_MAPPING
 from san.graphql import execute_gql, get_response_headers
 from san.query import get_gql_query, parse_dataset
 from san.transform import transform_timeseries_data_query_result
